Added an Adaptor Bundles to support HostTracker Porting
Per Colin's email:
https://lists.opendaylight.org/pipermail/discuss/2013-May/000195.html
about the plan for implementing the Dixon-Erickson Proposal item
for HostTracker:
https://wiki.opendaylight.org/view/D-E_Proposal:Host_Tracker_Plan

This commit brings an AdaptorBundle that should enable packetIn to
be adapted.  It provides a start on adapting IControllerService and
IOFSwitch in support of this.  Note, at this time, only the bits
needed to support the HostTracker Porting have been done.

Please also note, while this builds, and the bundle loads into
the OSGI container, DeviceManager will need to be OSGI-ized and
pushed to Nexus for it to be tested.

Because I do not know where the net-virt-platform guys are likely to want
the code put, I've put it in a 'staging' directory.

PatchSet2:
1. Cleaned up trailing whitespace
2. Switch from using locally defined NetYetImplementedException to
   java.lang.UnsupportedOperationException("Not implemented yet.")
3. Added a local .gitignore to protect against common badness like
   target/
Comment: lib/ is included intentionally until such time as the necessary
         binary artifacts are being pushed up to Nexus.  Until then, lib
         is a local maven repo.

PatchSet3:
1. Changed the pom file to work with https://git.opendaylight.org/gerrit/#/c/382/
and https://git.opendaylight.org/gerrit/#/c/386/
2. Removed the lib directory

PatchSet4:
1. Rebased to origin/master n

Change-Id: I04f8fbca6e66190046a9241f45bd5aff3413a757
Signed-off-by: Ed Warnicke <eaw@cisco.com>