Initial Maven POM build file
This POM builds the platform similarly to the way it
was being built with Ant, using the same dependencies
and versions, and unpacking and embedding them into
the resultant jar.  It also adds an OSGi MANIFEST.MF
so that the jar can be used within an OSGi runtime.
It currently exports a limited subset of the package
space, with the expectation that this will be expanded
over time as needed.

The build is currently excluding two test files from
running because they were failing at the time of this
patch.  Included in the POM is a note that these need
to be resolved.

Lastly I semi-arbitrarily picked the artifact name
'netvirtplatform-all', this code will need to evolve into
many more modules over time, and I expect this name to
change as well.

Change-Id: I551dafb141f93280fa503434f499f4709319ec32
Signed-off-by: David Erickson <daviderickson@cs.stanford.edu>